requirements.txt
playwright==1.44.0
python-telegram-bot==21.0
apscheduler==3.10.4
python-dotenv==1.0.1